adapter pattern
