+++ package.json
@@ -6,11 +6,14 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "node src/app.js",
-    "start:dev":"nodemon src/app.js"
+    "start:dev": "nodemon src/app.js"
   },
   "author": "Johao Lopez",
   "license": "ISC",
   "devDependencies": {
     "nodemon": "^3.0.1"
+  },
+  "dependencies": {
+    "uuid": "^9.0.1"
   }
 }
