+++ src/plugins/getuuid.plugin.js
@@ -1,0 +1,7 @@
+const { v4 } = require('uuid')
+
+const getUUID = () => v4()
+
+module.exports = {
+    getUUID
+}
